Contacts amended
- Bugs fixing.

+++ server/controllers/contact.controller.ts
@@ -183,7 +183,7 @@ function getContactDetails(req, res) {
   Contact.findOne({ userId: userId, _id: contactId })
     .then(resultContact => {
       const resultsFiltered = {
-        _id: resultContact._id,
+        id: resultContact._id,
         userId: resultContact.userId,
         givenName: resultContact.givenName,
         familyName: resultContact.familyName,
@@ -207,7 +207,7 @@ function getLists(req, res) {
     .then((result) => {
       const filteredResult = result.map((x) => {
         return {
-          listId: x._id,
+          id: x._id,
           listName: x.listName,
         };
       });
@@ -267,12 +267,12 @@ function deleteList(req, res) {
     return res.status(422).json({ message: 'Contact failed validation.' });
   }
 
-  return ContactListContact.find({ listId })
+  return ContactListContact.find({ userId, listId })
     .then (( result ) => {
       const contactId = result.map ((x => x.contactId));
       return Contact.deleteMany ({userId: userId, _id: contactId})
         .then (() => {
-          return ContactListContact.deleteMany({ listId: listId })
+          return ContactListContact.deleteMany({ userId, listId: listId })
             .then(() => {
               return ContactList.deleteMany({ refId: userId, _id: listId })
                 .then(() => {
@@ -298,10 +298,10 @@ function deleteList(req, res) {
 function editList(req, res) {
   const userId = req.userId;
   const listId = req.body.listId;
-  const UpdatedListName = req.body.UpdatedListName;
+  const updatedListName = req.body.updatedListName;
   // Validate
   if (typeof listId !== 'string' ||
-      typeof UpdatedListName !== 'string' ) {
+      typeof updatedListName !== 'string' ) {
     return res.status(422).json({ message: 'Request failed validation.' });
   }
 
@@ -312,7 +312,7 @@ function editList(req, res) {
   {
     $set:
     {
-      listName: UpdatedListName,
+      listName: updatedListName,
     }
   })
   .then(() => {
@@ -336,8 +336,8 @@ function getContactsWithLists(req, res) {
   const userId = req.userId;
   return Contact.find({ userId })
     .then((result) => {
-      const listIdArr = result.map((e => e._id));
-      return ContactListContact.find({ 'contactId': { '$in': listIdArr } })
+      const contactIdArr = result.map((e => e._id));
+      return ContactListContact.find({ userId: userId, 'contactId': { '$in': contactIdArr } })
         .then((result2) => {
           const filteredResult = result2.map((x) => {
             return {
@@ -371,7 +371,7 @@ function getContactsWithContactLists(req, res) {
   if (typeof listId !== 'string' ) {
     return res.status(422).json({ message: 'Request failed validation.' });
   }
-  return ContactListContact.find({ listId })
+  return ContactListContact.find({ userId, listId })
     .then((result) => {
       const contactId = result.map((x => x.contactId));
       return Contact.find({userId: userId, '_id': { '$in': contactId } })
